refactor(illustration-color-edit): split Streamlit app into one file per tab
Per project convention, app.py should stay thin and each tab lives in
its own file. Split the previous 425-line app.py into:

- app/app.py           74 lines, scaffolding + tab routing only
- app/common.py        shared UI helpers (swatches, badges, render_inline_svg,
                        cached_color_extract, fresh_mapper)
- app/tab_library.py
- app/tab_editor.py
- app/tab_global_map.py
- app/tab_batch.py
- app/tab_settings.py

Plus app/__init__.py to make the directory a proper package so each tab
module can import shared helpers via `from app.common import ...`.

README updated with the new layout. All 89 unit tests still pass.
