+++ illustration-color-edit/app/__init__.py
@@ -1,0 +1,1 @@
+"""Streamlit UI package — entry point in ``app.py``, one tab per ``tab_*.py``."""
